Adds notification support
It is now possible to send a notification for every backlight change.
The nofication is build for the dunst notification manager, which allows
showing a simple progress bar. This is why all notification values are
converted to relative ones and are passed as hint values to the
notification.
It is also possible to set the title of the notification.